requirements_test.txt
coverage==4.0a6
django-coverage-plugin==0.5
factory-boy==2.5.2
